LP configured correctly and new constraint, less coops added
LP is correct now, but usually only gets one cutvector (the other
is 0) and don't need all the coops. If it happens, solutionated
with a algorithm to decrease the number of cooperators.
New constraint added to avoid the comment before.

+++ src/RadioSystem/TransmissionCoef.cpp
@@ -24,7 +24,7 @@ float TransmissionCoef::getTransmissionTimeCoef() {
 }
 
 void TransmissionCoef::AddObservation(float txtime, int Npixels) {
-	float transmissiontcoeff= txtime/Npixels; // seconds/pixels (seconds/image)
+	double transmissiontcoeff= txtime/Npixels; // seconds/pixels (seconds/image)
 	if(training_samples <= TX_TRAINING_PERIOD){ //Training period: Arithmetic smoothing
 		training_samples++;
 		Ctcoef = ((training_samples-1)*Ctcoef + transmissiontcoeff)/training_samples;
